% Auto-generated: do not edit by hand
\name{dashPdfComponents}

\alias{dashPdfComponents}

\title{DashPdfComponents component}

\description{
PDF component using the Adobe Embed API.  Takes file url location and json annotations.
}

\usage{
dashPdfComponents(id=NULL, apiKey=NULL, currPdf=NULL, documentTable=NULL,
key=NULL, label=NULL, pdfRendered=NULL, showImpact=NULL,
toggle=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{apiKey}{Character. Api key for authentication}

\item{currPdf}{Character. URL to file that will be displayed}

\item{documentTable}{Unnamed list. Annotations that will appear as highlights}

\item{key}{Character. The ID used to identify this component in Dash callbacks.}

\item{label}{Character. A label that will be printed when this component is rendered.}

\item{pdfRendered}{Character. URL to file that will be displayed}

\item{showImpact}{Logical. }

\item{toggle}{Logical. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

